Update Traveling Salesman Recursion.sql

+++ Permutations, Combinations, Sequences and Random Numbers/Recursion Examples/Traveling Salesman Recursion.sql
@@ -19,7 +19,7 @@ CREATE TABLE #Routes
 (
 FromNode VARCHAR(30) NOT NULL,
 ToNode   VARCHAR(30) NOT NULL,
-Cost     INTEGER NOT NULL,
+Cost     MONEY NOT NULL,
 PRIMARY KEY (FromNode, ToNode)
 );
 GO
